+++ src/main/java/sws/songpin/domain/member/dto/request/MemberRequestDto.java
@@ -1,0 +1,4 @@
+package sws.songpin.domain.member.dto.request;
+
+public record MemberRequestDto() {
+}
