+++ src/main/java/sws/songpin/domain/member/entity/Status.java
@@ -1,0 +1,14 @@
+package sws.songpin.domain.member.entity;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+
+public enum Status {
+    ACTIVE,
+    DELETED
+    ;
+
+    @JsonCreator
+    public static Status from(String s) {
+        return Status.valueOf(s.toUpperCase());
+    }
+}
