--- src/main/java/sws/songpin/entity/follow/domain/Follow.java
@@ -1,34 +1,0 @@
-package sws.songpin.entity.follow.domain;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
-import lombok.*;
-import sws.songpin.entity.member.domain.Member;
-
-@Getter
-@Setter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-public class Follow {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "follow_id", updatable = false)
-    private Long followId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "follower_id", updatable = false)
-    @NotNull
-    private Member follower;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "following_id", updatable = false)
-    @NotNull
-    private Member following;
-
-    @Builder
-    public Follow(Long followId, Member follower, Member following) {
-        this.followId = followId;
-        this.follower = follower;
-        this.following = following;
-    }
-}
