--- src/main/java/sws/songpin/entity/member/domain/Status.java
@@ -1,14 +1,0 @@
-package sws.songpin.entity.member.domain;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-
-public enum Status {
-    ACTIVE,
-    DELETED
-    ;
-
-    @JsonCreator
-    public static Status from(String s) {
-        return Status.valueOf(s.toUpperCase());
-    }
-}
