--- src/main/java/sws/songpin/entity/pin/domain/Visibility.java
@@ -1,14 +1,0 @@
-package sws.songpin.entity.pin.domain;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-
-public enum Visibility {
-    PUBLIC,
-    PRIVATE
-    ;
-
-    @JsonCreator
-    public static Visibility from(String s) {
-        return Visibility.valueOf(s.toUpperCase());
-    }
-}
