--- src/main/java/sws/songpin/entity/song/repository/SongRepository.java
@@ -1,7 +1,0 @@
-package sws.songpin.entity.song.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import sws.songpin.entity.song.domain.Song;
-
-public interface SongRepository extends JpaRepository<Song, Long> {
-}
